+++ Mile.Helpers/Mile.Helpers.CppWinRT.cpp
@@ -11,10 +11,14 @@
 #include "Mile.Helpers.CppWinRT.h"
 
 #include <winrt/Windows.ApplicationModel.h>
+#include <winrt/Windows.ApplicationModel.Resources.Core.h>
+#include <winrt/Windows.Foundation.Collections.h>
 
 namespace winrt
 {
     using Windows::ApplicationModel::Package;
+    using Windows::ApplicationModel::Resources::Core::ResourceManager;
+    using Windows::ApplicationModel::Resources::Core::ResourceMap;
 }
 
 bool Mile::WinRT::IsPackagedMode()
@@ -33,4 +37,35 @@ bool Mile::WinRT::IsPackagedMode()
     }());
 
     return CachedResult;
+}
+
+winrt::hstring Mile::WinRT::GetLocalizedString(
+    winrt::hstring const& ResourcePath,
+    winrt::hstring const& FallbackString)
+{
+    try
+    {
+        winrt::ResourceMap CurrentResourceMap =
+            winrt::ResourceManager::Current().MainResourceMap();
+
+        if (CurrentResourceMap.HasKey(ResourcePath))
+        {
+            return CurrentResourceMap.Lookup(
+                ResourcePath).Candidates().GetAt(0).ValueAsString();
+        }
+        else
+        {
+            return FallbackString;
+        }
+    }
+    catch (...)
+    {
+        return FallbackString;
+    }
+}
+
+winrt::hstring Mile::WinRT::GetLocalizedString(
+    winrt::hstring const& ResourcePath)
+{
+    return Mile::WinRT::GetLocalizedString(ResourcePath, ResourcePath);
 }