+++ ReleaseNotes.md
@@ -30,3 +30,4 @@
 - Add Mile::WinRT::IsPackagedMode method.
 - Add MileIsCurrentProcessElevated function.
 - Add MileAllowNonClientDefaultDrawingForWindow function.
+- Add Mile::WinRT::GetLocalizedString method.